Implemented authentication

+++ backend/src/controllers/account/account.controller.ts
@@ -1,6 +1,7 @@
-import { Controller, Get, Post, Body } from '@nestjs/common';
+import { Controller, Get, Post, Body, Query } from '@nestjs/common';
 import { Account } from 'src/models/account';
 import { Config } from 'src/models/config';
+import { Global } from 'src/models/global';
 
 export class CreateAccountDto {
   name: string;
@@ -8,23 +9,32 @@ export class CreateAccountDto {
   password: string;
   broker: string;
   platform: string;
+  token: string;
 }
 
 export class DeleteAccountDto {
   name: string;
+  token: string;
 }
 
 export class StartAccountDto {
   name: string;
+  token: string;
 }
 
 export class StopAccountDto {
   name: string;
+  token: string;
 }
 
 export class ConfigAccountDto {
   name: string;
   configs: string[];
+  token: string;
+}
+
+export class ListAccountDto {
+  token: string;
 }
 
 @Controller('account')
@@ -49,9 +59,10 @@ export class AccountController {
   }
 
   @Get('list')
-  async listAccounts() {
+  async listAccounts(@Query() listAccountDto: ListAccountDto) {
     const account = new Account();
     try {
+      await Global.checkPermission(listAccountDto.token);
       const accounts = await account.getAccounts();
       return {
         status: 'success',
@@ -67,6 +78,7 @@ export class AccountController {
   async deleteAccount(@Body() deleteAccountDto: DeleteAccountDto) {
     const account = new Account();
     try {
+      await Global.checkPermission(deleteAccountDto.token);
       await account.deleteAccount(deleteAccountDto.name);
       return { status: 'success', message: 'Account deleted' };
     } catch (e) {
@@ -79,6 +91,7 @@ export class AccountController {
     const config = new Config();
     const account = new Account();
     try {
+      await Global.checkPermission(startAccountDto.token);
       await account.initAccount(startAccountDto.name);
       const configNames = await account.getConfigs(startAccountDto.name);
       for (const configName of configNames) {
@@ -99,6 +112,7 @@ export class AccountController {
   async stopAccount(@Body() stopAccountDto: StopAccountDto) {
     const account = new Account();
     try {
+      await Global.checkPermission(stopAccountDto.token);
       await account.stopAccount(stopAccountDto.name);
       return { status: 'success', message: 'Account stopped' };
     } catch (e) {
@@ -110,6 +124,7 @@ export class AccountController {
   async configAccount(@Body() configAccountDto: ConfigAccountDto) {
     const account = new Account();
     try {
+      await Global.checkPermission(configAccountDto.token);
       await account.updateConfigs(
         configAccountDto.name,
         configAccountDto.configs,
